pyproject.toml
[project]
name = "cw2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.11"
dependencies = [
    "accelerate>=1.12.0",
    "anthropic>=0.76.0",
    "boto3>=1.42.7",
    "codecarbon>=3.2.0",
    "datasets[torch]>=4.4.1",
    "devtools>=0.12.2",
    "hf-xet>=1.2.0",
    "huggingface-hub[cli,hf-xet]>=0.36.0",
    "ipykernel>=7.1.0",
    "ipython>=8.37.0",
    "joblib==1.4.2",
    "jupyter>=1.1.1",
    "lime>=0.2.0.1",
    "notebook>=7.5.0",
    "numpy>=2.2.6",
    "ollama>=0.6.1",
    "openai>=2.11.0",
    "openpyxl>=3.1.5",
    "pip>=25.3",
    "python-dotenv>=1.2.1",
    "scikit-learn==1.5.2",
    "shap>=0.49.1",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
    "transformers[torch]>=4.57.3",
]

[tool.uv.sources]
en-core-web-lg = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl" }
spacy-readability = { git = "https://github.com/Gems/spacy-readability" }
torch = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu126", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[tool.setuptools]
packages = ["src"]

[tool.uv]
package = true

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
